Add the demo conference
Add the initial work for Barcamp Huntsville. This demonstrates a very simple
(and silly) conference application that will playback howler monkeys to a
random channel in the conference when someone presses the '#' DTMF key. For
fun, the monkeys only whisper to the affected channel, such that they can
still (sort of) hear what the conference is discussing.